Display user datasets on main dashboard tab
We no longer display the news feed as it is not used. Instead, the main
dashboard page shows the user's datasets, and the tab link for 'My
Datasets' goes to the main dashboard.

The content of /dashboard/datasets is unchanged (so, the same as the
main dashboard) but it is not linked to anywhere.