Remove erroneously-added translations
These translations were automatically added to the .po files with the update_catalog command in setup.py. Most of them are translations that already exist in ckan or in various ckanexts, and overwriting them with empty strings in our .po files removes the existing translations, so only English is shown. Some of them have translated strings that are the same as the ones in ckan, and it's not necessary to include those in our ckanext as well.