Catch OverflowError when parsing POSIX date with dateutil
Here, we try to parse the date as a string formatted 'DD.MM.YYYY', so we can convert it into POSIX timestamp format. If we get a ParserError it was probably already in POSIX format and we return it as given. If it's a POSIX timestamp that is very large, however (e.g. -3342470400, which is Jan 31 1864), the error given is an OverflowError, so we need to catch that as well.