Fix format mapping for WORLDFILE
In practice, the media_type used to find the format in the mapping is split by '/' before being used. This means that although the media_type that we get from Geocat is 'JPEG/Worldfile (.jpg/.jgw)', we look up the format with '.jgw)', so that needs to be in the mapping.