repair temporal validator
after the the temporals are retrieved form the data, they need to be
cleaned: in some cases such as package_patch they come in as json strings
with iso format. So in that case a json.loads is necessary to convert
them to lists of dicts. Also the dates inside them have to be cleaned so that
they are stored in a uniform way