improve the date helper functions and implement test for timezone isoformat
the library isoformat is added as a dependecy. It's function `strftime` is also used for
formating dates before 1970 since the standard python's `strftime` can't do that

The helper functions are now documented. They always make sure that they can derive a datetime
from the given value. If they can they have the competence to return is either isodate
for storage in the database or in the datepicker format, that is configured in the
configuration file for display.

A test to store an isodate with a timezone is added.