Fix performance issue of archive interface before_update
The interface is called for a lot of instances (Package, PackageExtra,
Harvester, Organization, Groupd, etc.)

The method to find a unique new name for a deleted dataset is
expansive, since it has to query the datasets with similar names.
We need to make sure that this method is only called in appropriate
cases: if the instance is a Package, is of state deleted and is
not already one of these '_archived_...' copies of the datasets
that has been established on previous deletes of the dataset.