fix: Fix search form
The search form was overridden before we switched to the new frontend, and the overrides were not cleanly removed when we restored the templates to the ckan default.

With this change, the search_form.html snippet extends the ckan default snippet. The only change is to add the 'reset search' button. The button is styled inline as we don't have any other custom CSS in this extension.

This change restores the search result text (e.g. '25 datasets found'), the dropdown menu to select sort order, and the search button.