feat: Always substitute the correct url in footer
With the verification email, two links are added to the email_vars by ckanext.subscribe.send_verification_mail, after the email_vars have been set up by plugins implementing the ISubscribe interface (i.e., us). If the links are added to the footer before we can transform them to point at the frontend, and the footer is then added into the email text at the same time as substituting all the other email_vars into the text, we can't change them.

Instead, we let the footer keep its placeholders (e.g. {verification_link}) and add the footer to the email text. Then we substitute env_vars into the email text in a separate step. This means that we are using the correct links, which we have transformed.