feat: Remove manage link from verification email
This link does not contain an auth code, because the user has not yet verified their subscription and doesn't have one yet. In ckan, the link redirects to a page where the user can request a code be sent to them. On our frontend, we don't have such a page, so the link is useless.