feat: Handle resource data from c or from extras
The version of CKAN we are using stores template variables in the Pylons c object. This is deprecated (see https://github.com/ckan/ckan/issues/6769) and template variables should now be passed to the render function as extras.

This change has already been made by ckanext-xloader, so we need to be able to get template variables that are passed to the template in both ways.