feat: Only show button with link if the package group is 'energy'
Displaying the link inside a link like this is a bit messy. I tried setting a variable 'show_link' if the group name matches and then using that variable in an if block, but variables set inside loops in Jinja2 are reset outside the loop.

As this is a temporary proof-of-concept and will eventually be expanded to datasets in all groups, I think this is OK for now.