fix: If there is no value, save an empty list as the value
This fixes a bug when trying to save a dataset with no value for an extra field via the GUI: we want to not have the field at all, but we are left with a Missing object as the field's value, which causes an SQL error.

Save an empty list to replace the Missing object and prevent this error.