fix: Fix setting temporals in manage dataset form
If we were trying to set temporals from a dataset that previously had none, the value we were getting from the dataset data here was the string '[]'. This is truthy, so we were never getting the temporals data from the form. What we want to see here is an empty list, which is falsey.