chore: Remove overridden user_info block
This was overridden so we could show the user's login status in the user_info block: are they allowed to log in, or are they locked out?

This template change should have been in ckanext-password-policy all along and now it has been moved there, so we can delete this here.