feat: Update search form reset-search button
Groups (and group-like things) have an index function that leads to the index page. Datasets (and dataset-like things) have a search function that leads to their index page. In the search form, we don't actually know if we are searching for ggroups or datasets, so it's hard to know where to link to in the reset-search button. However, we have a custom search for organizations, and we are not using any other group-like entities here, so we can just treat the type 'group' as a special case and assume everything else we might search for is a dataset-like entity.