fix: Don't pass in url_defaults, set variables in view func
The CKAN organization blueprint is the same as the group one, but passes in these url default values everywhere. We don't want them in our blueprint as it messes up the xml_upload view func. We are only using the index view func here, so set those variables inside it directly.