fix: Update check for organization controller
HierarchyDisplay checks whether before_dataset_search is being called from the organization controller by checking whether the g object is registered, and whether it has a list of fields in it. But then what it actually uses is g.group_dict. If we are being called from, say, the user controller, we might have g.fields but not g.group_dict. Let's just check for g.group_dict in the first place.