fix: Don't include groups when getting organisation to localise it
This might prevent a recursive error where, in ckan.lib.dictization.model_dictize, the function group_list calls group_list_dictize if include_groups=True, but then group_list_dictize also calls group_list, and this keeps going. I haven't been able to reproduce this locally so I'm not sure this is a fix.

In any case, when we call this helper method, we never need the organization's sub-organizations, so it should not break anything.