fix: Fix renaming datasets when they are deleted
This method is before the package is deleted, so its state is still 'active', not deleted.

Furthermore, the method is only called when deleting a package from the dataset (ckan.logic.action.delete.package_delete), so we already know the entity is a Package and don't need to check this.