fix: Remove ignore_missing validator for resource license
This validator stops all further validation for a field if the key is missing, or if the key is there and the value is None. (The docstring is misleading here.) This is the reverse of what we want, which is that if the 'license' value is None, we should go on to ogdch_license_required and copy the value from the resource 'rights' field.

All our harvesters should add the license key even if the value is None - we don't want to ignore that the key is missing, because that would indicate an actual error.