fix: Fix validation of language field
This validator only allows saving values in the list of languages if they belong to the list of choices for the field, as defined in the dataset schema, i.e. 'de', 'en', 'fr' or 'en'. There is no point checking for the EU vocabulary URI and then not doing anything with it.